+++ src/plugins/application.py
@@ -204,7 +204,7 @@ async def updateApplication(request: Request, response: Response, authorization:
     applicationid = form["applicationid"]
     message = form["message"]
 
-    if applicationid < 0:
+    if int(applicationid) < 0:
         response.status_code = 404
         return {"error": True, "descriptor": ml.tr(request, "application_not_found")}
 
@@ -330,7 +330,7 @@ async def updateApplicationStatus(request: Request, response: Response, authoriz
     if int(status) in STATUS.keys():
         statustxt = STATUS[int(status)]
 
-    if applicationid < 0:
+    if int(applicationid) < 0:
         response.status_code = 404
         return {"error": True, "descriptor": ml.tr(request, "application_not_found")}
 
@@ -425,7 +425,7 @@ async def getApplication(request: Request, response: Response, applicationid: in
     conn = newconn()
     cur = conn.cursor()
 
-    if applicationid < 0:
+    if int(applicationid) < 0:
         response.status_code = 404
         return {"error": True, "descriptor": ml.tr(request, "application_not_found")}